fix: 포인트 값 타입 Long으로 변경, rabbitMQ 로직 추가

+++ src/main/java/com/service/point/domain/entity/PointUsageHistory.java
@@ -19,13 +19,13 @@ public class PointUsageHistory {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long pointUsageHistoryId;
     private LocalDate pointUsageHistoryDate;
-    private Integer pointUsageAmount;
+    private Long pointUsageAmount;
     @ManyToOne
     @JoinColumn(name = "pointUsageTypeId")
     private PointUsageType pointUsageType;
     private long clientId;
 
-    public PointUsageHistory(Integer pointUsageAmount, PointUsageType pointUsageType,
+    public PointUsageHistory(Long pointUsageAmount, PointUsageType pointUsageType,
         long clientId) {
         this.pointUsageHistoryDate=LocalDate.now();
         this.pointUsageAmount = pointUsageAmount;
