+++ src/test/java/com/service/point/dto/response/TotalPointAmountResponseDtoTest.java
@@ -17,15 +17,15 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
         // Then
         assertNotNull(dto);
-        assertEquals(null, dto.getTotalPoint()); // Check default value for Integer (null)
-        assertEquals(null, dto.getPointAccumulationRate()); // Check default value for Integer (null)
+        assertEquals(null, dto.getTotalPoint()); // Check default value for Long (null)
+        assertEquals(null, dto.getPointAccumulationRate()); // Check default value for Long (null)
     }
 
     @Test
      void testSetterAndGetter() {
         // Given
-        Integer totalPoint = 1000;
-        Integer pointAccumulationRate = 5;
+        Long totalPoint = 1000;
+        Long pointAccumulationRate = 5;
 
         // When
         TotalPointAmountResponseDto dto = new TotalPointAmountResponseDto();
